+++ src/istate.cpp
@@ -1,7 +1,7 @@
 // Functions for state interface.
 #include "istate.hh"
 
-IState::IState(IActor* act, Trie<float>* config) : actor(act), behavioralConfig(config) 
+IState::IState(IActor* act, Trie<double>* config) : actor(act), behavioralConfig(config) 
 {
 	DEBUG_LOG("ISTATE", "Constructing IState.");
 	dispatcher = new Dispatcher();
@@ -18,7 +18,7 @@ void IState::setActor(IActor* actor)
 	this->actor = actor;
 }
 
-void IState::setConfig(Trie<float>* config)
+void IState::setConfig(Trie<double>* config)
 {
 	behavioralConfig = config;
 }
